+++ client/src/pages/FindId/FindId.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import { FiArrowLeft } from 'react-icons/fi';
 import ModalContainer from '../../components/login/LoginModal';
-import { EmailForm, FindIdContainer, GoToLogin, InputBox, ModalTitle, SendButton, WhatFindContainer } from './Style';
+import { EmailForm, FindIdContainer, InputBox, SendButton, WhatFindContainer } from './Style';
 
 const FindId = () => {
   const [showAuth, setShowAuth] = useState<boolean>(false);
@@ -22,11 +21,7 @@ const FindId = () => {
   };
 
   return (
-    <ModalContainer>
-      <GoToLogin to="/login">
-        <FiArrowLeft size={20} />
-      </GoToLogin>
-      <ModalTitle>아이디 찾기</ModalTitle>
+    <ModalContainer title="아이디 찾기">
       <WhatFindContainer>
         <Link className="find-id" to="/find_id">
           아이디 찾기
